pyproject.toml
[project]
name = "jquants-mcp"
version = "0.1.0"
description = "MCP server for J-Quants API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jquants-api-client>=2.0.0",
    "mcp[cli]>=1.26.0",
    "pandas>=2.2.0",
    "pydantic>=2.12.5",
]

[project.scripts]
jquants-mcp = "jquants_mcp.__main__:main"

[dependency-groups]
dev = ["ruff>=0.9.6"]
